chore: reorganize and restructure code
