--- src/authentication/index.ts
@@ -1,8 +1,0 @@
-export { Toolbox, type ToolboxProps } from './components/Toolbox';
-export {
-  useAuthentication,
-  useOrfiumProducts,
-  useOrganizations,
-  useTopBarUtilitySection,
-} from './context';
-export * from './types';
