--- src/hooks/data/index.ts
@@ -1,2 +1,0 @@
-export * from './useDataPagination';
-export * from './useData';
