+++ src/hooks/useAuthentication.ts
@@ -1,0 +1,4 @@
+import { useContext } from 'react';
+import { AuthenticationContext } from '../contexts/authentication';
+
+export const useAuthentication = () => useContext(AuthenticationContext);
