+++ src/hooks/useOrfiumProducts.ts
@@ -1,0 +1,4 @@
+import { useContext } from 'react';
+import { OrfiumProductsContext } from '../contexts/orfium-products';
+
+export const useOrfiumProducts = () => useContext(OrfiumProductsContext);
