+++ src/hooks/useOrganizations.ts
@@ -1,0 +1,4 @@
+import { useContext } from 'react';
+import { OrganizationsContext } from '../contexts/organizations';
+
+export const useOrganizations = () => useContext(OrganizationsContext);
