+++ src/routing/index.tsx
@@ -2,4 +2,11 @@
 // That is why we keep react-router-dom as dependency in package.json
 
 export { BrowserRouter, Redirect, Route } from 'react-router-dom';
-export * from './Routing';
+export {
+  Authorization,
+  FallbackPath,
+  generateRoutes,
+  RouteComponentProps,
+  RouteItem,
+  RoutingStructure,
+} from './Routing';
