+++ src/ui/index.ts
@@ -1,3 +1,3 @@
-export * from './consts';
-export * from './Navigation';
-export * from './TopBar';
+export { DEFAULT_NAVBAR_HEIGHT } from './consts';
+export { Navigation, type MenuItem, type NavigationProps } from './Navigation';
+export { TopBar, type TopBarProps } from './TopBar';
